bringing up the latest borg codebase
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@7279 f5854215-dd47-0410-b2c4-cdd35faa7885

+++ vision/borg/scan.cpp
@@ -11,7 +11,7 @@ int main(int, char **)
   //cam->takePhoto(IMAGESIZE_640_480, raster);
   cam->startImageStream();
   uint8_t *raster = cam->savePhoto();
-  printf("raster = %x\n", raster);
+  printf("raster = %x\n", (unsigned)raster);
   if (raster)
     delete[] raster;
   cam->stopImageStream();
